+++ docs/conf.py
@@ -34,6 +34,7 @@ linkcheck_timeout = 5
 extensions = [
     'sphinx.ext.mathjax',
     'sphinx.ext.doctest',
+    'sphinxcontrib.programoutput',
 ]
 
 # Add any paths that contain templates here, relative to this directory.