+++ rasa_nlu/extractors/entity_synonyms.py
@@ -6,7 +6,6 @@ from __future__ import unicode_literals
 import os
 import warnings
 
-import six
 from builtins import str
 from typing import Any
 from typing import Dict
@@ -20,14 +19,19 @@ from rasa_nlu.training_data import Message
 from rasa_nlu.training_data import TrainingData
 from rasa_nlu.utils import write_json_to_file
 
+ENTITY_SYNONYMS_FILE_NAME = "entity_synonyms.json"
+
 
 class EntitySynonymMapper(EntityExtractor):
     name = "ner_synonyms"
 
     provides = ["entities"]
 
-    def __init__(self, synonyms=None):
+    def __init__(self, component_config=None, synonyms=None):
         # type: (Optional[Dict[Text, Text]]) -> None
+
+        super(EntitySynonymMapper, self).__init__(component_config)
+
         self.synonyms = synonyms if synonyms else {}
 
     def train(self, training_data, config, **kwargs):
@@ -50,51 +54,62 @@ class EntitySynonymMapper(EntityExtractor):
         message.set("entities", updated_entities, add_to_output=True)
 
     def persist(self, model_dir):
-        # type: (Text) -> Dict[Text, Any]
+        # type: (Text) -> Optional[Dict[Text, Any]]
 
         if self.synonyms:
             entity_synonyms_file = os.path.join(model_dir,
-                                                "entity_synonyms.json")
-            write_json_to_file(entity_synonyms_file, self.synonyms, separators=(',', ': '))
+                                                ENTITY_SYNONYMS_FILE_NAME)
+            write_json_to_file(entity_synonyms_file, self.synonyms,
+                               separators=(',', ': '))
 
-            return {"entity_synonyms": "entity_synonyms.json"}
-        else:
-            return {"entity_synonyms": None}
+        return {"synonyms_file": ENTITY_SYNONYMS_FILE_NAME}
 
     @classmethod
-    def load(cls, model_dir, model_metadata, cached_component, **kwargs):
-        # type: (Text, Metadata, Optional[EntitySynonymMapper], **Any) -> EntitySynonymMapper
+    def load(cls,
+             model_dir=None,  # type: Optional[Text]
+             model_metadata=None,  # type: Optional[Metadata]
+             cached_component=None,  # type: Optional[EntitySynonymMapper]
+             **kwargs  # type: **Any
+             ):
+        # type: (...) -> EntitySynonymMapper
 
-        if model_dir and model_metadata.get("entity_synonyms"):
-            entity_synonyms_file = os.path.join(model_dir, model_metadata.get("entity_synonyms"))
-            if os.path.isfile(entity_synonyms_file):
-                synonyms = utils.read_json_file(entity_synonyms_file)
-                return EntitySynonymMapper(synonyms)
-            else:
-                warnings.warn("Failed to load synonyms file from '{}'".format(entity_synonyms_file))
-        return EntitySynonymMapper()
+        meta = model_metadata.for_component(cls.name)
+        file_name = meta.get("synonyms_file", ENTITY_SYNONYMS_FILE_NAME)
+        entity_synonyms_file = os.path.join(model_dir, file_name)
+
+        if os.path.isfile(entity_synonyms_file):
+            synonyms = utils.read_json_file(entity_synonyms_file)
+        else:
+            synonyms = None
+            warnings.warn("Failed to load synonyms file from '{}'"
+                          "".format(entity_synonyms_file))
+
+        return EntitySynonymMapper(meta, synonyms)
 
     def replace_synonyms(self, entities):
         for entity in entities:
-            entity_value = str(entity["value"])  # need to wrap in `str` to handle e.g. entity values of type int
+            # need to wrap in `str` to handle e.g. entity values of type int
+            entity_value = str(entity["value"])
             if entity_value.lower() in self.synonyms:
                 entity["value"] = self.synonyms[entity_value.lower()]
                 self.add_processor_name(entity)
 
     def add_entities_if_synonyms(self, entity_a, entity_b):
         if entity_b is not None:
-            original = entity_a if isinstance(entity_a, six.text_type) else six.text_type(entity_a)
-
-            replacement = entity_b if isinstance(entity_b, six.text_type) else six.text_type(entity_b)
+            original = utils.as_text_type(entity_a)
+            replacement = utils.as_text_type(entity_b)
 
             if original != replacement:
                 original = original.lower()
-                if original in self.synonyms and self.synonyms[original] != replacement:
-                    warnings.warn("Found conflicting synonym definitions for {}. Overwriting "
-                                  "target {} with {}. Check your training data and remove conflicting "
-                                  "synonym definitions to prevent this from happening.".format(repr(original),
-                                                                                               repr(self.synonyms[
-                                                                                                        original]),
-                                                                                               repr(replacement)))
+                if (original in self.synonyms
+                        and self.synonyms[original] != replacement):
+                    warnings.warn("Found conflicting synonym definitions "
+                                  "for {}. Overwriting target {} with {}. "
+                                  "Check your training data and remove "
+                                  "conflicting synonym definitions to "
+                                  "prevent this from happening."
+                                  "".format(repr(original),
+                                            repr(self.synonyms[original]),
+                                            repr(replacement)))
 
                 self.synonyms[original] = replacement