+++ rasa_nlu/server.py
@@ -1,57 +1,85 @@
-from __future__ import unicode_literals
-from __future__ import print_function
-from __future__ import division
 from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
+from __future__ import unicode_literals
 
 import argparse
 import logging
-import os
-import six
 from functools import wraps
 
 import simplejson
+import six
 from builtins import str
-
 from klein import Klein
 from twisted.internet import reactor, threads
 from twisted.internet.defer import inlineCallbacks, returnValue
 
-from rasa_nlu.config import RasaNLUConfig
-from rasa_nlu.data_router import DataRouter, InvalidProjectError, \
-    AlreadyTrainingError
+from rasa_nlu import utils, config
+from rasa_nlu.config import RasaNLUModelConfig
+from rasa_nlu.data_router import (
+    DataRouter, InvalidProjectError,
+    AlreadyTrainingError)
 from rasa_nlu.train import TrainingException
-from rasa_nlu.version import __version__
 from rasa_nlu.utils import json_to_string
+from rasa_nlu.version import __version__
 
 logger = logging.getLogger(__name__)
 
 
-def create_argparser():
+def create_argument_parser():
     parser = argparse.ArgumentParser(description='parse incoming text')
-    parser.add_argument('-c', '--config',
-                        help="config file, all the command line options can "
-                             "also be passed via a (json-formatted) config "
-                             "file. NB command line args take precedence")
+
     parser.add_argument('-e', '--emulate',
                         choices=['wit', 'luis', 'dialogflow'],
                         help='which service to emulate (default: None i.e. use '
                              'simple built in format)')
-    parser.add_argument('-l', '--language',
-                        choices=['de', 'en'],
-                        help="model and data language")
-    parser.add_argument('-p', '--path',
-                        help="path where project files will be saved")
-    parser.add_argument('--pipeline',
-                        help="The pipeline to use. Either a pipeline template "
-                             "name or a list of components separated by comma")
     parser.add_argument('-P', '--port',
                         type=int,
+                        default=5000,
                         help='port on which to run server')
     parser.add_argument('-t', '--token',
                         help="auth token. If set, reject requests which don't "
                              "provide this token as a query parameter")
     parser.add_argument('-w', '--write',
                         help='file where logs will be saved')
+    parser.add_argument('--path',
+                        required=True,
+                        help="working directory of the server. Models are"
+                             "loaded from this directory and trained models "
+                             "will be saved here.")
+    parser.add_argument('--cors',
+                        nargs="*",
+                        help='List of domain patterns from where CORS '
+                             '(cross-origin resource sharing) calls are '
+                             'allowed. The default value is `[]` which '
+                             'forbids all CORS requests.')
+
+    parser.add_argument('--max_training_processes',
+                        type=int,
+                        default=1,
+                        help='Number of processes used to handle training '
+                             'requests. Increasing this value will have a '
+                             'great impact on memory usage. It is '
+                             'recommended to keep the default value.')
+    parser.add_argument('--num_threads',
+                        type=int,
+                        default=1,
+                        help='Number of parallel threads to use for '
+                             'handling parse requests.')
+    parser.add_argument('--response_log',
+                        help='Directory where logs will be saved '
+                             '(containing queries and responses).'
+                             'If set to ``null`` logging will be disabled.')
+    parser.add_argument('--storage',
+                        help='Set the remote location where models are stored. '
+                             'E.g. on AWS. If nothing is configured, the '
+                             'server will only serve the models that are '
+                             'on disk in the configured `path`.')
+    parser.add_argument('-c', '--config',
+                        help="Default model configuration file used for "
+                             "training.")
+
+    utils.add_logging_option_arguments(parser)
 
     return parser
 
@@ -66,9 +94,9 @@ def check_cors(f):
         origin = request.getHeader('Origin')
 
         if origin:
-            if '*' in self.config['cors_origins']:
+            if '*' in self.cors_origins:
                 request.setHeader('Access-Control-Allow-Origin', '*')
-            elif origin in self.config['cors_origins']:
+            elif origin in self.cors_origins:
                 request.setHeader('Access-Control-Allow-Origin', origin)
             else:
                 request.setResponseCode(403)
@@ -93,7 +121,7 @@ def requires_auth(f):
             token = request.args.get(b'token', [b''])[0].decode("utf8")
         else:
             token = str(request.args.get('token', [''])[0])
-        if self.config['token'] is None or token == self.config['token']:
+        if self.access_token is None or token == self.access_token:
             return f(*args, **kwargs)
         request.setResponseCode(401)
         return 'unauthorized'
@@ -101,25 +129,74 @@ def requires_auth(f):
     return decorated
 
 
+def decode_parameters(request):
+    """Make sure all the parameters have the same encoding.
+
+    Ensures  py2 / py3 compatibility."""
+    return {
+        key.decode('utf-8', 'strict'): value[0].decode('utf-8', 'strict')
+        for key, value in request.args.items()}
+
+
+def parameter_or_default(request, name, default=None):
+    """Return a parameters value if part of the request, or the default."""
+
+    request_params = decode_parameters(request)
+    return request_params.get(name, default)
+
+
+def dump_to_data_file(data):
+    if isinstance(data, six.string_types):
+        data_string = data
+    else:
+        data_string = utils.json_to_string(data)
+
+    return utils.create_temporary_file(data_string, "_training_data")
+
+
+def is_yaml_request(request):
+    return "yml" in next(
+            iter(request.requestHeaders.getRawHeaders("Content-Type", [])), "")
+
+
 class RasaNLU(object):
     """Class representing Rasa NLU http server"""
 
     app = Klein()
 
-    def __init__(self, config, component_builder=None, testing=False):
-        logging.basicConfig(filename=config['log_file'],
-                            level=config['log_level'])
-        logging.captureWarnings(True)
-        logger.debug("Configuration: " + config.view())
+    def __init__(self,
+                 data_router,
+                 loglevel='INFO',
+                 logfile=None,
+                 num_threads=1,
+                 token=None,
+                 cors_origins=None,
+                 testing=False,
+                 default_config_path=None):
 
-        logger.debug("Creating a new data router")
-        self.config = config
-        self.data_router = self._create_data_router(config, component_builder)
+        self._configure_logging(loglevel, logfile)
+
+        self.default_model_config = self._load_default_config(
+                default_config_path)
+
+        self.data_router = data_router
         self._testing = testing
-        reactor.suggestThreadPoolSize(config['num_threads'] * 5)
+        self.cors_origins = cors_origins if cors_origins else ["*"]
+        self.access_token = token
+        reactor.suggestThreadPoolSize(num_threads * 5)
 
-    def _create_data_router(self, config, component_builder):
-        return DataRouter(config, component_builder)
+    @staticmethod
+    def _load_default_config(path):
+        if path:
+            return config.load(path).as_dict()
+        else:
+            return {}
+
+    @staticmethod
+    def _configure_logging(loglevel, logfile):
+        logging.basicConfig(filename=logfile,
+                            level=loglevel)
+        logging.captureWarnings(True)
 
     @app.route("/", methods=['GET', 'OPTIONS'])
     @check_cors
@@ -131,11 +208,10 @@ class RasaNLU(object):
     @requires_auth
     @check_cors
     @inlineCallbacks
-    def parse_get(self, request):
+    def parse(self, request):
         request.setHeader('Content-Type', 'application/json')
         if request.method.decode('utf-8', 'strict') == 'GET':
-            request_params = {key.decode('utf-8', 'strict'): value[0].decode('utf-8', 'strict')
-                              for key, value in request.args.items()}
+            request_params = decode_parameters(request)
         else:
             request_params = simplejson.loads(
                     request.content.read().decode('utf-8', 'strict'))
@@ -145,7 +221,8 @@ class RasaNLU(object):
 
         if 'q' not in request_params:
             request.setResponseCode(404)
-            dumped = json_to_string({"error": "Invalid parse parameter specified"})
+            dumped = json_to_string(
+                    {"error": "Invalid parse parameter specified"})
             returnValue(dumped)
         else:
             data = self.data_router.extract(request_params)
@@ -177,8 +254,12 @@ class RasaNLU(object):
     def rasaconfig(self, request):
         """Returns the in-memory configuration of the Rasa server"""
 
+        # DEPRECATED: I don't think there is a use case for this endpoint
+        # anymore - when training a new model, the user should always post
+        # the configuration as part of the request instead of relying on
+        # the servers config.
         request.setHeader('Content-Type', 'application/json')
-        return json_to_string(self.config.as_dict())
+        return json_to_string(self.default_model_config)
 
     @app.route("/status", methods=['GET', 'OPTIONS'])
     @requires_auth
@@ -192,16 +273,32 @@ class RasaNLU(object):
     @check_cors
     @inlineCallbacks
     def train(self, request):
-        data_string = request.content.read().decode('utf-8', 'strict')
-        kwargs = {key.decode('utf-8', 'strict'): value[0].decode('utf-8', 'strict')
-                  for key, value in request.args.items()}
+        project = parameter_or_default(request, "project", default=None)
+
+        request_content = request.content.read().decode('utf-8', 'strict')
+
+        if is_yaml_request(request):
+            # assumes the user submitted a model configuration with a data
+            # parameter attached to it
+            model_config = utils.read_yaml(request_content)
+            data = model_config.get("data")
+        else:
+            # assumes the caller just provided training data without config
+            # this will use the default model config the server
+            # was started with
+            model_config = self.default_model_config
+            data = request_content
+
+        data_file = dump_to_data_file(data)
+
         request.setHeader('Content-Type', 'application/json')
 
         try:
             request.setResponseCode(200)
             response = yield self.data_router.start_train_process(
-                    data_string, kwargs)
-            returnValue(json_to_string({'info': 'new model trained: {}'.format(response)}))
+                    data_file, project, RasaNLUModelConfig(model_config))
+            returnValue(json_to_string({'info': 'new model trained: {}'
+                                                ''.format(response)}))
         except AlreadyTrainingError as e:
             request.setResponseCode(403)
             returnValue(json_to_string({"error": "{}".format(e)}))
@@ -246,7 +343,7 @@ class RasaNLU(object):
         try:
             request.setResponseCode(200)
             response = self.data_router.unload_model(
-                params.get('project', RasaNLUConfig.DEFAULT_PROJECT_NAME),
+                params.get('project', RasaNLUModelConfig.DEFAULT_PROJECT_NAME),
                 params.get('model')
             )
             return simplejson.dumps(response)
@@ -258,12 +355,22 @@ class RasaNLU(object):
 
 if __name__ == '__main__':
     # Running as standalone python application
-    arg_parser = create_argparser()
-    cmdline_args = {key: val
-                    for key, val in list(vars(arg_parser.parse_args()).items())
-                    if val is not None}
-    rasa_nlu_config = RasaNLUConfig(
-            cmdline_args.get("config"), os.environ, cmdline_args)
-    rasa = RasaNLU(rasa_nlu_config)
-    logger.info('Started http server on port %s' % rasa_nlu_config['port'])
-    rasa.app.run('0.0.0.0', rasa_nlu_config['port'])
+    cmdline_args = create_argument_parser().parse_args()
+
+    utils.configure_colored_logging(cmdline_args.loglevel)
+
+    router = DataRouter(cmdline_args.path,
+                        cmdline_args.max_training_processes,
+                        cmdline_args.response_log)
+    rasa = RasaNLU(
+            router,
+            cmdline_args.loglevel,
+            cmdline_args.write,
+            cmdline_args.num_threads,
+            cmdline_args.token,
+            cmdline_args.cors,
+            default_config_path=cmdline_args.config
+    )
+
+    logger.info('Started http server on port %s' % cmdline_args.port)
+    rasa.app.run('0.0.0.0', cmdline_args.port)