+++ rasa_nlu/tokenizers/whitespace_tokenizer.py
@@ -1,16 +1,13 @@
-from __future__ import unicode_literals
-from __future__ import print_function
-from __future__ import division
 from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
+from __future__ import unicode_literals
 
-from typing import Any
-from typing import Dict
-from typing import List
-from typing import Text
+from typing import Any, List, Text
 
-from rasa_nlu.config import RasaNLUConfig
-from rasa_nlu.tokenizers import Tokenizer, Token
 from rasa_nlu.components import Component
+from rasa_nlu.config import RasaNLUModelConfig
+from rasa_nlu.tokenizers import Tokenizer, Token
 from rasa_nlu.training_data import Message
 from rasa_nlu.training_data import TrainingData
 
@@ -21,7 +18,7 @@ class WhitespaceTokenizer(Tokenizer, Component):
     provides = ["tokens"]
 
     def train(self, training_data, config, **kwargs):
-        # type: (TrainingData, RasaNLUConfig, **Any) -> None
+        # type: (TrainingData, RasaNLUModelConfig, **Any) -> None
 
         for example in training_data.training_examples:
             example.set("tokens", self.tokenize(example.text))