+++ rasa_nlu/train.py
@@ -1,22 +1,21 @@
-from __future__ import unicode_literals
-from __future__ import print_function
-from __future__ import division
 from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
+from __future__ import unicode_literals
 
 import argparse
 import logging
-import os
 
 import typing
+from typing import Optional, Any
 from typing import Text
 from typing import Tuple
-from typing import Optional
 
+from rasa_nlu import utils, config
 from rasa_nlu.components import ComponentBuilder
+from rasa_nlu.config import RasaNLUModelConfig
 from rasa_nlu.model import Interpreter
 from rasa_nlu.model import Trainer
-
-from rasa_nlu.config import RasaNLUConfig
 from rasa_nlu.training_data import load_data
 
 logger = logging.getLogger(__name__)
@@ -25,26 +24,47 @@ if typing.TYPE_CHECKING:
     from rasa_nlu.persistor import Persistor
 
 
-def create_argparser():
+def create_argument_parser():
     parser = argparse.ArgumentParser(
             description='train a custom language parser')
 
-    parser.add_argument('-p', '--pipeline', default=None,
-                        help="Pipeline to use for the message processing.")
-    parser.add_argument('-o', '--path', default=None,
+    parser.add_argument('-o', '--path',
+                        default=None,
                         help="Path where model files will be saved")
-    parser.add_argument('-d', '--data', default=None,
-                        help="File containing training data")
-    parser.add_argument('-c', '--config', required=True,
+
+    parser.add_argument('-d', '--data',
+                        default=None,
+                        required=True,
+                        help="Location of the training data. For JSON and "
+                             "markdown data, this can either be a single file "
+                             "or a directory containing multiple training "
+                             "data files.")
+
+    parser.add_argument('-c', '--config',
+                        required=True,
                         help="Rasa NLU configuration file")
-    parser.add_argument('-l', '--language', default=None, choices=['de', 'en'],
-                        help="Model and data language")
-    parser.add_argument('-t', '--num_threads', default=None, type=int,
+
+    parser.add_argument('-t', '--num_threads',
+                        default=None,
+                        type=int,
                         help="Number of threads to use during model training")
+
+    parser.add_argument('--project',
+                        default=None,
+                        help="Project this model belongs to.")
+
     parser.add_argument('--fixed_model_name',
                         help="If present, a model will always be persisted "
                              "in the specified directory instead of creating "
                              "a folder like 'model_20171020-160213'")
+
+    parser.add_argument('--storage',
+                        help='Set the remote location where models are stored. '
+                             'E.g. on AWS. If nothing is configured, the '
+                             'server will only serve the models that are '
+                             'on disk in the configured `path`.')
+
+    utils.add_logging_option_arguments(parser)
     return parser
 
 
@@ -65,33 +85,33 @@ class TrainingException(Exception):
         return self.message
 
 
-def create_persistor(config):
-    # type: (RasaNLUConfig) -> Optional[Persistor]
+def create_persistor(persistor):
+    # type: (Optional[Text]) -> Optional[Persistor]
     """Create a remote persistor to store the model if configured."""
 
-    if config.get("storage") is not None:
+    if persistor is not None:
         from rasa_nlu.persistor import get_persistor
-        return get_persistor(config)
+        return get_persistor(persistor)
     else:
         return None
 
 
-def init():  # pragma: no cover
-    # type: () -> RasaNLUConfig
-    """Combines passed arguments to create rasa NLU config."""
-
-    parser = create_argparser()
-    args = parser.parse_args()
-    config = RasaNLUConfig(args.config, os.environ, vars(args))
-    return config
-
-
-def do_train_in_worker(config):
-    # type: (RasaNLUConfig) -> Text
+def do_train_in_worker(config,  # type: RasaNLUModelConfig
+                       data,  # type: Text
+                       path,  # type: Text
+                       project=None,  # type: Optional[Text]
+                       fixed_model_name=None,  # type: Optional[Text]
+                       storage=None,  # type: Text
+                       component_builder=None
+                       # type: Optional[ComponentBuilder]
+                       ):
+    # type: (...) -> Text
     """Loads the trainer and the data and runs the training in a worker."""
 
     try:
-        _, _, persisted_path = do_train(config)
+        _, _, persisted_path = do_train(config, data, path, project,
+                                        fixed_model_name, storage,
+                                        component_builder)
         return persisted_path
     except Exception as e:
         logger.exception("Failed to train project '{}'.".format(
@@ -99,8 +119,14 @@ def do_train_in_worker(config):
         raise TrainingException(config.get("project"), e)
 
 
-def do_train(config,  # type: RasaNLUConfig
-             component_builder=None  # type: Optional[ComponentBuilder]
+def do_train(cfg,  # type: RasaNLUModelConfig
+             data,  # type: Text
+             path=None,  # type: Text
+             project=None,  # type: Optional[Text]
+             fixed_model_name=None,  # type: Optional[Text]
+             storage=None,  # type: Text
+             component_builder=None,  # type: Optional[ComponentBuilder]
+             **kwargs   # type: Any
              ):
     # type: (...) -> Tuple[Trainer, Interpreter, Text]
     """Loads the trainer and the data and runs the training of the model."""
@@ -108,19 +134,32 @@ def do_train(config,  # type: RasaNLUConfig
     # Ensure we are training a model that we can save in the end
     # WARN: there is still a race condition if a model with the same name is
     # trained in another subprocess
-    trainer = Trainer(config, component_builder)
-    persistor = create_persistor(config)
-    training_data = load_data(config['data'], config['language'])
-    interpreter = trainer.train(training_data)
-    persisted_path = trainer.persist(config['path'], persistor,
-                                     config['project'],
-                                     config['fixed_model_name'])
+    trainer = Trainer(cfg, component_builder)
+    persistor = create_persistor(storage)
+    training_data = load_data(data, cfg.language)
+    interpreter = trainer.train(training_data, **kwargs)
+
+    if path:
+        persisted_path = trainer.persist(path,
+                                         persistor,
+                                         project,
+                                         fixed_model_name)
+    else:
+        persisted_path = None
+
     return trainer, interpreter, persisted_path
 
 
 if __name__ == '__main__':
-    config = init()
-    logging.basicConfig(level=config['log_level'])
+    cmdline_args = create_argument_parser().parse_args()
 
-    do_train(config)
+    utils.configure_colored_logging(cmdline_args.loglevel)
+
+    do_train(config.load(cmdline_args.config),
+             cmdline_args.data,
+             cmdline_args.path,
+             cmdline_args.project,
+             cmdline_args.fixed_model_name,
+             cmdline_args.storage,
+             num_threads=cmdline_args.num_threads)
     logger.info("Finished training")